+++ NEWS.md
@@ -1,5 +1,8 @@
 # marquee (development version)
 
+* Fixed a bug in bullet placement that affected tight lists with multiple spans 
+  (#18)
+
 # marquee 0.1.0
 
 * Initial CRAN submission.