+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Fixed a bug in bullet placement that affected tight lists with multiple spans 
   (#18)
+* code spans gains a slight horizontal padding to let the background breathe a 
+  bit. Currently padding around spans doesn't affect shaping (i.e. it doesn't 
+  move text further from it's neighbors)
 
 # marquee 0.1.0