+++ NEWS.md
@@ -4,7 +4,12 @@
   (#18)
 * code spans gains a slight horizontal padding to let the background breathe a 
   bit. Currently padding around spans doesn't affect shaping (i.e. it doesn't 
-  move text further from it's neighbors)
+  move text further from it's neighbors).
+* Better adherence to margin collapsing rules of CSS. Any background or border
+  will now prevent further collapsing
+* Add `force_body_margin` argument to enforce that the body margin is not 
+  influenced by collapsing (allowing you to turn it off completely). This 
+  setting is turned on for `geom_marquee()` and `element_marquee()` (#23)
 
 # marquee 0.1.0